requirements.txt
# Data Collection
pandas==1.5.3
yfinance==0.2.18
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2

# Data Storage
pymongo==4.5.0
hdfs==2.7.0
kafka-python==2.0.2

# Data Processing
pyspark==3.4.1
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
ta==0.10.2  # Technical Analysis library

# Machine Learning
tensorflow==2.13.0
keras==2.13.1
statsmodels==0.14.0
xgboost==1.7.6
lightgbm==4.0.0

# Visualization
flask==2.2.5
dash==2.13.0
plotly==5.16.1
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
pytest==7.4.0
jupyter==1.0.0